Updates tsp-openapi3 to emit main.tsp even when formatter fails (#3794)
This updates the `tsp-openapi3` CLI so that it emits `main.tsp` even
when the formatter detects an issue. The CLI will still throw the error
after it is done emitting the file.

The `convertOpenAPI3Document` function used by the website does not
throw on formatter errors - the playground will highlight any TSP errors
already.

Playground example with bad emission:

![image](https://github.com/microsoft/typespec/assets/14189820/cbddf049-5928-4969-b1fa-cbf98ea95bc0)

---------

Co-authored-by: Christopher Radek <Christopher.Radek@microsoft.com>