Normalize mocking in mgc.tests (#3807)
We were all doing mocking slightly differently. I created a common
helper to mock the plugin and provide some functionality to pass in
behaviors that we want to change on a per test basis. I will make a
follow up PR to update mgc.cm to follow the same pattern.