[http-server-javascript] Fix an annoying bug with router path matching. (#4115)
This PR fixes a bug where path parameters in the middle of a path would
fail to match because they incorrectly removed the leading "/" from the
`rest` of the path segment after matching. The "/" trailing a parameter
was never included as part of the path but was discarded by a call to
`String.split`. The logic in this PR is more robust, always using
indices and slices.

---------

Co-authored-by: Will Temple <will@wtemple.net>