implement array as header and query, and format different type to string (#4213)
Fix https://github.com/microsoft/typespec/issues/4205
Fixes https://github.com/microsoft/typespec/issues/3973

Problem to solve:
When any type except string (e.g. int, TimeSpan and so on) as
header/query, it is not correct covert to correct format, and when there
is an array as header/query, it does not serialize correct in connection
format (such as cvs, pipe and so on).

- call TypeFormatters.ConvertToString to covert this type to correct
format
- when it is an array, we need to convert to string in correct
collection format.