Fix the type cache with extensibility (#4427)
Fixes #4428 

This PR also fixes the incorrect caching system for `CreateCSharpType`.
The `CreateCSharpType` is the entry point, therefore the type cache
should work here - and all the invocations in the `CreateCSharpTypeCore`
should be verified that this input is not in the cache, therefore the
cache should never be accessed again once we go into this method.