[http-client-python] fix ci for http-client-python (#4541)
Step of ci to publish cadl-ranch report
[fails](https://dev.azure.com/azure-sdk/internal/_build/results?buildId=4161916&view=logs&j=9b5a911b-af88-5d05-2d29-054ea7ec9492&t=32de1fc2-ab38-529d-8066-ce0987282b63)
because test step actually doesn't succeed:
https://dev.azure.com/azure-sdk/internal/_build/results?buildId=4161916&view=logs&j=9b5a911b-af88-5d05-2d29-054ea7ec9492&t=0bd06dfe-81a0-5fb5-18dc-3c1cf5f385b4.
We are in a hurry to release first version of http-client-python to
unblock release of typepspec-python, so skip the step to publish
cadl-ranch report for now. We have another repo to refresh the
cadl-ranch report so it doesn't have influence.