[@typespec/http-specs] Add withKeys Logic for server-test and payload/content-negotiation service (#4579)
During the migration of `cadl-ranch-specs` package from `cadl-ranch`
repository to `typespec` repository (as `http-spec` package), the
`payload\content-negotiation` service was not migrated. Because, it was
based on keys which was not implemented for the `server-test` scenarios.

In this PR, the implementation is completed and the
`payload\content-negotiation` service has been added back.

1. I considered adding `|` to `withKeys` instead of adding
`withServiceKeys`. But, it was creating build issues with incompatible
types. The new `withServiceKeys` implementation is simpler and cleaner
approach.
2. The previous `cadl-ranch` package has [2
scenarios](https://github.com/Azure/cadl-ranch/blob/main/packages/cadl-ranch-specs/http/payload/content-negotiation/mockapi.ts):
    a. Scenarios.Payload_ContentNegotiation_SameBody
    b. Scenarios.Payload_ContentNegotiation_DifferentBody
The same 2 scenarios are added in the new file also.
3. The `server-test` of the 2 new scenarios is successful. Also, the
`validate-scenarios` and `validate-mock-api` have been successful for
the new scenarios. Here are the associated snapshots:

![Screenshot 2024-09-30
130737](https://github.com/user-attachments/assets/c2b52759-0537-4c17-8fbd-faa23c184859)

![Screenshot 2024-09-30
130754](https://github.com/user-attachments/assets/9b08ebee-52d4-4280-a706-e499616e57a6)

Please review and approve this PR. Thanks