[@typespec/http-specs] Add payload/multipart service (#4594)
During the migration of cadl-ranch-specs package from cadl-ranch
repository to typespec repository (as http-spec package), the
payload/multipart service was not migrated. (This is the last pending
one)

In this PR, the implementation is completed and the routes service has
been added back. I have already run the `pnpm test:e2e` and it is
successful.

Please review and approve this PR. Thanks