[@typespec/spector] - Modify checks for headers & parameters (#4751)
As a continuation of https://github.com/microsoft/typespec/pull/4725,
while working on removing the handler for the scenarios in the
`typespec-azure` repository, I found some scenarios need added fix in
the `spector` library.

This PR consists of 2 changes:

1. The header keys must be changed to lowercase before reading and
comparing.
2. If the headers and parameters are array values, then we need to
handle them seperately.

I have tested this change with all the scenarios in both the `typespec`
and `typespec-azure` repositories. Once this change is merged, I can
open the PR for the scenario changes in `typespec-azure` repository.

Please review and approve this PR. Thanks