[@typespec/http-specs] Added Spec Summary (#4768)
During the modification of Scenarios in the past weeks, this
spec-summary was not added because the scenarios were undergoing
modifications. Now, as the migration is complete, this spec summary
could be generated.

I have compared the generated file with the file in the `cadl-ranch`
summary and confirmed that it is exactly the same except for one change.
In the `cadl-ranch` repository, the `type/model/templated` scenarios
have been removed recently. Those scenarios are still present in this
`typespec` repository and part of this spec summary.

Please review and approve the PR. Thanks