[http-client-csharp] fix: use inputmodel name as default name in serialization provider (#4823)
This PR fixes an issue where the name of the serialization provider for
a model was being built incorrectly when the model is renamed and the
renamed model is also further customized. It addresses this issue by
ensuring the default name of the serialization provider is the name of
the input model it is constructed from.

fixes: https://github.com/microsoft/typespec/issues/4793