[@typespec/spector] - Enable Multiple Scenario Paths (#4836)
Currently we could execute the `serve`/`server-test`/... command such
as:

`pnpm .\cmd\cli.mjs serve ..\http-specs\specs`

i.e we could provide only one path of scenarios to execute. But, since
we separated the scenarios to azure and non-azure specific cases to 2
different packages, we need a way to execute commands such as:

`pnpm .\cmd\cli.mjs serve ..\http-specs\specs
..\..\..\typespec-azure\packages\azure-http-specs\specs\`

The same procedure must be followed in all other commands such as
`validate-scenarios`, `generate-scenarios-summary`, etc. Since we have
to do it for all the commands and we need to change in all the places,
it is much simpler to do it in one location where we copy the files to
one common temp location and execute from there.

Also, in this PR, I have removed the `private: true` option for all the
5 packages to start the rollout process.

Please review and approve the PR. Thanks