[@typespec/spector] - Handle and report failed scenarios at the end (#4872)
In the `server-test` scenario, if there is an error, then the script
immediately reports the error and comes out of the execution. But, the
service generator has not implemented all the scenarios yet. So, If all
the scenarios are executed, then the script might fail even before
executing the intended script. So, the correct approach would be to
report at the end.

Please review and approve the PR. Thanks