[@typespec/spector] [@typespec/spec-coverage-sdk] - Handle Upload Scenario Manifest  (#4920)
Earlier, in the `upload-scenario-manifest` script, I had handled the
multiple paths in a different way (one by one). After working with the
`cadl-ranch-dashboard`, the correct approach is to generate the manifest
file in the same way as coverage file. This PR performs this task.

Here is a snapshot of the modified manifest file:


![image](https://github.com/user-attachments/assets/cf752245-452a-424f-b1ed-86b2965103a0)

Please review and approve the PR. Thanks