Adds an emitter shell and sets up the tests to get options (#5033)
### Description

This PR sets up the flow to use the GraphQL emitter by providing an
interface for the various options that the GraphQL emitter will use
eventually. It also sets up test-hosts to work with these options. The
actual schema emitter doesn't really do anything other than emit "Hello
World" as it did previously, but the options get pass through as
confirmed by the test case.

Going forward, we can change the code in schema-emitter.ts to setup it
up for GraphQL using `navigateProgram`. We need to add diagnostics in
the emitter lib definition, but that can be done in a separate PR. The
next PR will have the outer layer of the GraphQL emitter setup to deal
with multiple schemas similar to multiple services in the OAI emitter.

### Testing
Run the tests and see that they pass.

---------

Co-authored-by: swatikumar <swatikumar@pinterest.com>