adds initial Open API 3.1 schema emitter (#5035)
Related to #5009 and #5010

This PR starts the work of getting an Open API 3.1 emitter in the
`@typespec/openapi3` package.

There are a few changes introduced in this PR:
1. Added the `OpenAPIDocument3_1` and related types. The biggest change
here is updating the Open API schema to be based on Json Schema 2020-12.
The Json Schema types don't currently exist in `@typespec/json-schema`
so I added them to `@typespec/openapi3` for now while iterating. Can
move to `@typespec/json-schema` now or later if we want.
2. Created the OpenAPISchemaEmitter3_1 class to target Open API 3.1
schemas. Currently completely standalone from Json Schema and Open API
3.0 schema emitters.
3. Updated nearly all tests to now run against both Open API 3.x
versions. Intentionally left out the enum tests and schema examples
since those will be updated shortly and be different between the two
versions.
4. Added support for `type: "null"` in the 3.1 schema emitter. I
squeezed this change in because the changes involved were actually
really small, and this way TypeScript was happy we were emitting 3.1
schemas that matched the defined interface (no `nullable`). I copied the
behavior that the Json Schema emitter uses for handling nulls.
5. Updated `exclusiveMinimum/Maximum` to be a number value instead of a
boolean to match the Open API 3.1 spec.

Note to reviewers: Hide whitespace - otherwise the test changes will
look monstrous.

---------

Co-authored-by: Christopher Radek <Christopher.Radek@microsoft.com>