http-client-java, eng, fix script for release pipeline (#5127)
fix https://github.com/microsoft/typespec/issues/5046

- Copy `cadl-ranch-coverage-java-standard.json` to the location that
will be uploaded by shared script
- Move cadl-ranch test to "UnitTests" block ("GenerationChecks" is the
diff on re-generated code -- it won't run in publish)
- Update publish.yml to disable NuGet, and enable release to "public"

Note that this coverage is actually not via unbranded. It is Azure.
We should be able to switch the cases to unbranded, when we got
clientcore lib on maven.