http-client-java, namespace from client.tsp (#5051)
link https://github.com/Azure/autorest.java/issues/2968?reload=1

current limitation:
- operation group be in same package of client
- subclient be in same package of client

mostly this due to the fact that ctor of the client is package private,
so it won't work if client and its builder or client of accessor method
are in different packages

---

At this PR, unbranded would still have `.models` in package for models.
We can tweak this later (by setting `modelsSubPackage` in JavaSetting to
empty, when unbranded).

---

tested and it be backward compatible, if `namespace` is provided in
emitter option
https://github.com/Azure/azure-sdk-for-java/pull/43055/files