[http-server-js] Ignore unspeakable properties (#5188)
Closes #5185

This adds some extra handling for properties with names like `_`. In the
current implementation, our property name parser will parse an empty
list of name segments that ultimately combines to an empty string in any
case convention.

This PR addresses this problem by declaring certain properties
"unspeakable". A property name is "unspeakable" if its parsed
representation would be an empty string. Unspeakable properties are
removed from the representation of a Model. If the value of an
"unspeakable" property is demanded, the emitter will check that the
property has a literal valued type and convert it to a value rather than
trying to access the property in the subject value, or issue an error
diagnostic if it is not a literal type (`@statusCode _: 200` is okay,
`@statusCode _: uint16` is not). This effectively makes the types of
unspeakable properties "associated constants" of the Model, rather than
variable properties of the model. Unspeakable properties are ignored in
the RTTI code generator, so that unspeakable properties will never be
used to try to introspect the type of a value.

Further, the change uncovered a problem with model differentiation logic
in which a property was not always guaranteed to exist on a type before
its value was acquired, resulting in a type checking error. The code
generation for checking properties that may not exist on object
instances has been improved from `subject.prop === expected` to `"prop"
in subject && subject.prop === expected`.

This fixes a problem with the generation of REST specs that commonly use
`_` as the name of the statusCode property.

---------

Co-authored-by: Will Temple <will@wtemple.net>