build: strict use pnpm instead of npm
What
===

Use `pnpm run` instead of `npm run` in the `pnpm run build` when in root.

I have also changed `npx` to `pnpm dlx`.

I haven't touched anything that looks like a Windows environment, such as PowerShell,
because I'm not sure what it is supposed to do.

`vsce package` add option `--no-dependencies`.
vsce package support npm and yarn, but not support pnpm.
When it is executed, `pnpm install` should be finished, so it should be fine.

Why
===

When using corepack,
it is strictly decided which package manager to use,
and if npm is run even though pnpm is specified to be used,
as in this project, an error occurs.

Other environments where pnpm is installed but npm is not would also cause problems.

Therefore, I could not complete the `pnpm run build`.

I will try to unify pnpm as much as possible,
because there is no happiness at all if I call npm in an environment that already depends on pnpm.

Since pnpm is already assumed, there is no disadvantage at all if invoking pnpm.