Remove unnecessary lazy for InputLibrary (#5232)
Before this PR, `InputLibrary` depends on `Instance.Configuration`, and
`Instance` is the current `CodeModelPlugin` instance being constructed.

https://github.com/microsoft/typespec/blob/b9e5648e0bc106c47f66b9146562f2ba1341b62a/packages/http-client-csharp/generator/Microsoft.Generator.CSharp/src/CodeModelPlugin.cs#L48
This introduces circular dependency of `InputLibrary` on `Instance`, but
`InputLirabry` can just depend on `Configuration`, which is in the same
class, which breaks the circular dependency.
Therefore, `lazy` is unnecessary in this case.