[http-server-javascript] Two small correctness fixes (#5253)
This fixes two issues that showed up in the widget and petstore REST
examples.

- Required query parameters were being detected as missing if their
value was not exactly `null`. This PR changes this to a falsiness test.
- Checking literal range-constrained properties in type differentiation
could try to use a property without proving its presence. This PR adds
an `in` check to that logic so that range-constrained properties have
the same guard that literal-valued properties do.

---------

Co-authored-by: Will Temple <will@wtemple.net>