[@typespec/spector] Update start server scenario paths (#5309)
With my earlier PRs, I have already moved all the `scenarioPath`
parameters to `scenarioPaths` parameter i.e. made that parameter from a
single value to an array of values.

But, this start script was missed earlier. But, the Go SDK generator
uses this script and need this to be updated. This PR acoomplishes that.

Please review and approve this PR. Thanks