`typespec` repository - Upload Manifest file to the new storage account (#5336)
In the `cadl-ranch` repository, during the publish process, the CI
publishes the manifest file to the `azuresdkcadlranch` storage account.
You can find the code
[here](https://github.com/Azure/cadl-ranch/blob/main/.azure-pipelines/publish.yml#L35)

With the new Spec Dashboard, the `typespec` repository CI has to publish
the manifest file to the `typespec` storage account. I have tested this
code in CI in my branch. You can find the link
[here](https://dev.azure.com/azure-sdk/internal/_build/results?buildId=4401569&view=logs&j=feeac6a1-95f5-5e4f-b973-2aeaa4cef844).

Please review and approve the PR. Thanks