publish vscode artifact and display in pr comment for try-it (#5206)
Fix https://github.com/microsoft/typespec/issues/5155

When there is any changes/PR about typespec-vscode, we want to provide
an extension to try. So we will publish `typespec-vscode-0.XXX.0.vsix`
as pipeline artifact, and will show `VSCode Extension` with the download
url in the try-it comment. Customer can download this extension file to
try.

---------

Co-authored-by: Mark Cowlishaw <markcowl@microsoft.com>