Migrate cadl-ranch specs to http-specs and azure-http-specs (#5357)
This is the non-azure part of
https://github.com/Azure/azure-sdk-for-net/issues/47325

We split previous cadl-ranch tests into two parts: pure Typespec:
[http-specs](https://www.npmjs.com/package/@typespec/http-specs), and
Azure Typespec:
[azure-http-specs](https://www.npmjs.com/package/@azure-tools/azure-http-specs).

This PR is to remove the old cadl-ranch and enable the tests from these
two new ones. Then generate coverage report for both but just upload the
unbranded one to dashboard.

The coverage can be successfully uploaded in this test build.
https://dev.azure.com/azure-sdk/internal/_build/results?buildId=4408920&view=results

---------

Co-authored-by: Pan Shao <pashao@microsoft.com>