playground - support array options (#5310)
This PR updates the playground to handle array emitter options.

Currently this mainly supports options that are an array of enums - such
as the new openapi3 versions option.

If an option accepts an array of non-enums, then currently it will
display a single input field, effectively restricting usage to a single
element. As new array options come up we can decide how we want to
display them (e.g. show a '+' sign and create a new element?)

Screenshot of the `Openapi Versions` option showing both choices as
selectable:

![image](https://github.com/user-attachments/assets/955d2ea0-2ec0-435c-8805-7efb5f50b5c4)

Co-authored-by: Christopher Radek <Christopher.Radek@microsoft.com>