[core] Fix enum-driven visibility consistency issues when projections are used (#5525)
- Changed `getLifecycleVisibilityEnum` to honor projections when
`program` is a `ProjectedProgram`.
- Taught `Projector` to project `EnumValue` values only, so that they
remain consistent with projected `Enum` and `EnumMember` references. If
the projection of the underlying `EnumMember` does not result in another
`EnumMember` (i.e. it projects to some other type, which I think should
be rare, but possible according to the type system), the projector gives
up and returns the original enum member.

A more complex solution would try to project all values, but this could
damage the scalar hierarchy and there is an open issue to consider
removing projections in favor of mutators in the future.

Closes #5450
Closes #5461

---------

Co-authored-by: Will Temple <will@wtemple.net>