[http-client-csharp]: Fix explode not being properly set for operation params (#5396)
This PR fixes an issue where the `explode` flag for a operation
parameter was not being properly set when uri templates were specified.
This is fixed now by passing through the value of `explode` from TCGC.

There are several follow ups that are still needed to fully support URI
templates and path expansion.

fixes: https://github.com/microsoft/typespec/issues/5134
regen preview for autorest:
https://github.com/Azure/azure-sdk-for-net/pull/47775