Fix the xml doc issue when the parameter's actual name is not the same as their variable name (#5615)
Fixes https://github.com/microsoft/typespec/issues/5565
Fixes https://github.com/microsoft/typespec/issues/5567
Fixes https://github.com/microsoft/typespec/issues/5568

The root cause of the issue here is:
the `ParameterProvider` has a `Name` property, but it is actually the
name of its input parameter, not its actual name in the code which we
will never know until we actually write them into the code (because of
the scope stuff and name collisions)

Therefore, this PR added a scope version of `WriteXmlDocs` and allows
the parameters to be declared in the xml doc.
For a method, the parameters first appear in the xml docs, and then the
method signature, therefore allowing the xml doc part to declare the
parameter actually makes sense - since our code writer could only go
forwards, no backwards.
With this change, if there is xml doc, when the code writer writes the
xml doc for a parameter, if the parameter has a "not declared"
`CodeWriterDeclaration`, it will declare it with the current scope
(which is basically the same scope of the method), and then when it
writes the method signature, it will find those parameters have been
declared and it will just use the name, therefore everything works fine.
If there is no xml doc, when the code writer writes the method
signature, it will find those parameters are not declared yet, and
declares them. This is the same logic before this change.

To achieve this, this PR has to change the start tag and end tag in the
`XmlDocProvider` to be `FormattableString`s. This allows us to utilize
the ability of `CodeWriter` to automatically resolve the declaration of
the parameter inside the tag.