Promote `state-accessor` functions from experimental to supported (#5699)
These reusable state accessors were added in
3eed29834600d02c6a1ab3dd039188edf0fb2951 under the `experimental`
namespace. They are now considered ready to be promoted out to the
`@typespec/compiler` package exports (and to have their `unsafe_`
prefixes removed).

In this commit, we move `state-accessor` into `utils` and add it to the
module index so that its functions can be imported from
`@typespec/compiler/utils`.

For usages inside `@typespec/compiler`, we are replacing the wrapper
functions in `lib/utils` with direct imports of the newly-exposed
accessor functions. Because the wrapper functions previously created a
symbol with the `TypeSpec.` prefix when given a string, we preserve this
behavior by introducing the simple `createStateSymbol` function in files
that previously used the wrapper functions.

`createStateSymbol` could be trivially refactored into a shared utility
function rather than duplicated, but here we are simply following the
pattern already established where there are already several copies of
this function. Indeed, while one of these was exported in
050139d2e272bb0b1f4d02010276a554c0b4ddc4, that export was never used.